Kotlin/Common Buffer read and write hexadecimal unsigned long
A benchmark was created to make sure the transition from Java utilities
to custom algorithm for counting the hexadecimal width of a Long did not
regress. These are the results of running on my Windows laptop with a
Core i7-7700HQ Intel CPU and 8 GB of ram.

Previous performance:

Benchmark                           (width)   Mode  Cnt   Score   Error   Units
WriteHexadecimalBenchmark.writeHex        1  thrpt    5  25.467 ± 0.634  ops/us
WriteHexadecimalBenchmark.writeHex        2  thrpt    5  25.562 ± 0.353  ops/us
WriteHexadecimalBenchmark.writeHex        3  thrpt    5  25.516 ± 0.512  ops/us
WriteHexadecimalBenchmark.writeHex        4  thrpt    5  25.143 ± 0.689  ops/us
WriteHexadecimalBenchmark.writeHex        5  thrpt    5  25.191 ± 0.268  ops/us
WriteHexadecimalBenchmark.writeHex        6  thrpt    5  25.180 ± 0.292  ops/us
WriteHexadecimalBenchmark.writeHex        7  thrpt    5  25.138 ± 0.548  ops/us
WriteHexadecimalBenchmark.writeHex        8  thrpt    5  24.141 ± 0.239  ops/us
WriteHexadecimalBenchmark.writeHex        9  thrpt    5  24.215 ± 0.122  ops/us
WriteHexadecimalBenchmark.writeHex       10  thrpt    5  24.297 ± 0.176  ops/us
WriteHexadecimalBenchmark.writeHex       11  thrpt    5  24.149 ± 0.943  ops/us
WriteHexadecimalBenchmark.writeHex       12  thrpt    5  23.532 ± 2.501  ops/us
WriteHexadecimalBenchmark.writeHex       13  thrpt    5  24.034 ± 0.166  ops/us
WriteHexadecimalBenchmark.writeHex       14  thrpt    5  24.080 ± 0.388  ops/us
WriteHexadecimalBenchmark.writeHex       15  thrpt    5  23.979 ± 0.414  ops/us
WriteHexadecimalBenchmark.writeHex       16  thrpt    5  23.535 ± 0.116  ops/us

Current performance:

Benchmark                           (width)   Mode  Cnt   Score   Error   Units
WriteHexadecimalBenchmark.writeHex        1  thrpt    5  23.525 ± 1.686  ops/us
WriteHexadecimalBenchmark.writeHex        2  thrpt    5  23.740 ± 0.488  ops/us
WriteHexadecimalBenchmark.writeHex        3  thrpt    5  23.785 ± 0.258  ops/us
WriteHexadecimalBenchmark.writeHex        4  thrpt    5  22.887 ± 2.215  ops/us
WriteHexadecimalBenchmark.writeHex        5  thrpt    5  23.140 ± 0.746  ops/us
WriteHexadecimalBenchmark.writeHex        6  thrpt    5  23.266 ± 0.317  ops/us
WriteHexadecimalBenchmark.writeHex        7  thrpt    5  23.236 ± 0.542  ops/us
WriteHexadecimalBenchmark.writeHex        8  thrpt    5  22.477 ± 0.427  ops/us
WriteHexadecimalBenchmark.writeHex        9  thrpt    5  22.458 ± 0.270  ops/us
WriteHexadecimalBenchmark.writeHex       10  thrpt    5  22.278 ± 0.388  ops/us
WriteHexadecimalBenchmark.writeHex       11  thrpt    5  22.418 ± 0.441  ops/us
WriteHexadecimalBenchmark.writeHex       12  thrpt    5  22.183 ± 0.641  ops/us
WriteHexadecimalBenchmark.writeHex       13  thrpt    5  22.266 ± 0.573  ops/us
WriteHexadecimalBenchmark.writeHex       14  thrpt    5  22.183 ± 0.525  ops/us
WriteHexadecimalBenchmark.writeHex       15  thrpt    5  22.207 ± 0.331  ops/us
WriteHexadecimalBenchmark.writeHex       16  thrpt    5  21.696 ± 0.209  ops/us
